requirements.txt
flask
uwsgi
github3.py
requests
